move fixtures.py to conftest.py
